Breakout DNSCache into zeroconf.cache
